deprecate old namespace
