+++ pypsa/plot/accessor.py
@@ -3,7 +3,9 @@
 import functools
 from typing import TYPE_CHECKING, Any
 
-from pypsa.plot.maps import plot
+from deprecation import deprecated
+
+from pypsa.plot.maps import explore, iplot, plot
 
 if TYPE_CHECKING:
     from pypsa import Network
@@ -23,7 +25,25 @@ class PlotAccessor:
         """Initialize the statistics accessor."""
         self.n = n  # TODO rename
 
+    @deprecated(
+        details="Use `n.plot.map()` as a drop-in replacement instead.",
+    )
     @functools.wraps(plot)
     def __call__(self, *args: Any, **kwargs: Any) -> Any:  # type: ignore
         """Legacy plot method."""
         return plot(self.n, *args, **kwargs)
+
+    @functools.wraps(plot)
+    def map(self, *args: Any, **kwargs: Any) -> Any:
+        """Plot method."""
+        return plot(self.n, *args, **kwargs)
+
+    @functools.wraps(iplot)
+    def iplot(self, *args: Any, **kwargs: Any) -> Any:
+        """Interactive plot method."""
+        return iplot(self.n, *args, **kwargs)
+
+    @functools.wraps(explore)
+    def explore(self, *args: Any, **kwargs: Any) -> Any:
+        """Interactive map plot method."""
+        return explore(self.n, *args, **kwargs)
